Also remove BeanValidationMessages
