+++ src/main/resources/META-INF/rewrite/jakarta-ee-10.yml
@@ -32,11 +32,12 @@ recipeList:
   - org.openrewrite.java.migrate.jakarta.RemovalsServletJakarta10
   - org.openrewrite.java.migrate.jakarta.DeprecatedCDIAPIsRemoved40
   - org.openrewrite.java.migrate.BeanDiscovery
-  - org.openrewrite.java.migrate.jakarta.BeanValidationMessages
   - org.openrewrite.java.migrate.jakarta.UpdateAnnotationAttributeJavaxToJakarta:
       signature: "@jakarta.ejb..*"
   - org.openrewrite.java.migrate.jakarta.UpdateAnnotationAttributeJavaxToJakarta:
       signature: "@jakarta.jms..*"
+  - org.openrewrite.java.migrate.jakarta.UpdateAnnotationAttributeJavaxToJakarta:
+      signature: "@jakarta.validation.constraints..*"
   - org.openrewrite.java.migrate.jakarta.JavaxBeansXmlToJakartaBeansXml
   - org.openrewrite.java.migrate.jakarta.JavaxEjbJarXmlToJakartaEjbJarXml
   - org.openrewrite.java.migrate.jakarta.JavaxBeanValidationXmlToJakartaBeanValidationXml
